подключения сервисов в контейнер сервисов
