+++ Services/WebStore.ServiceHosting/WebStore.ServiceHosting.csproj
@@ -8,4 +8,11 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\Common\WebStore.Domain\WebStore.Domain.csproj" />
+    <ProjectReference Include="..\Webstore.DAL\Webstore.DAL.csproj" />
+    <ProjectReference Include="..\Webstore.Interfaces\Webstore.Interfaces.csproj" />
+    <ProjectReference Include="..\Webstore.Services\Webstore.Services.csproj" />
+  </ItemGroup>
+
 </Project>
