+++ Services/WebStore.ServiceHosting/appsettings.json
@@ -1,4 +1,7 @@
 {
+  "ConnectionStrings": {
+    "Default": "Data Source=(localdb)\\MSSQLLocalDB;Initial Catalog=WebStore-DB"
+  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
